requirements.txt
pyTelegramBotAPI
requests
beautifulsoup4
psutil
pyautogui
pillow
speedtest-cli
rich
